requirements.txt
djangorestframework
asn1crypto==0.24.0
astroid==1.6.2
attrs==18.1.0
Automat==0.7.0
autopep8==1.3.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
Django==2.1.2
django-bootstrap-form==3.4
django-constance==2.2.0
django-crontab==0.7.1
django-extensions==2.1.2
django-picklefield==1.1.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
parsel==1.5.0
persisting-theory==0.2.1
proxy-requests==0.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==1.8.3
pyOpenSSL==18.0.0
python-scrapyd-api==2.1.2
pytz==2018.5
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.1
scrapyd==1.2.0
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
wrapt==1.10.11
zope.interface==4.5.0
